GeoAlert-Pipeline/kafka-producer/requirements.txt
kafka-python==2.0.2
beautifulsoup4==4.12.3
requests==2.32.3
schedule==1.1.0
spacy==3.7.6
